refactor : 음식 개수 검증 메서드 OrderMenu클래스로 이동
